put tests in right file
